[api-login-register-jwt2.0] login register hooks
